画像処理機能を切り出す
